Move the iiwa lcm handlers into the common library.
